tried to make wait for enter work
