+++ src/RandomHandler.cpp
@@ -3,19 +3,16 @@
 #include <thread>
 #include <chrono>
 #include <mutex>
+#include <iostream>
+#include <atomic>
+#include <condition_variable>
 #include "../include/RandomHandler.hpp"
 
+std::condition_variable cv;
+std::atomic<bool> running(true);
+std::mutex mtx;
 std::random_device rd;
 std::mt19937 gen(rd());
-
-struct RandomTransactionData
-{
-    BankAccount &account;
-    int clientId;
-    int accountNumber;
-    int amount;
-    std::mutex &bankMutex;
-};
 
 int getRandomNumber(int min, int max)
 {
@@ -40,40 +37,34 @@ int getRandomAccount()
 
 void depositRandom(RandomTransactionData data)
 {
-    while (true)
+
     {
-        {
-            std::lock_guard<std::mutex> lock(data.bankMutex);
-            data.account.deposit(data.amount);
-            std::cout << "Client " << data.clientId << " deposited " << data.amount << " to account " << data.accountNumber << std::endl;
-        }
-        std::this_thread::sleep_for(std::chrono::milliseconds(6000));
+        std::lock_guard<std::mutex> lock(data.bankMutex);
+        data.account.deposit(data.amount);
+        std::cout << "Client " << data.clientId << " deposited " << data.amount << " to account " << data.accountNumber << std::endl;
     }
+    std::this_thread::sleep_for(std::chrono::milliseconds(2000));
 }
 
 void withdrawRandom(RandomTransactionData data)
 {
-    while (true)
+
     {
-        {
-            std::lock_guard<std::mutex> lock(data.bankMutex);
-            data.account.withdraw(data.amount);
-            std::cout << "Client " << data.clientId << " withdrew " << data.amount << " from account " << data.accountNumber << std::endl;
-        }
-        std::this_thread::sleep_for(std::chrono::milliseconds(10000));
+        std::lock_guard<std::mutex> lock(data.bankMutex);
+        data.account.withdraw(data.amount);
+        std::cout << "Client " << data.clientId << " withdrew " << data.amount << " from account " << data.accountNumber << std::endl;
     }
+    std::this_thread::sleep_for(std::chrono::milliseconds(3000));
 }
 
 void showBalance(RandomTransactionData data)
 {
-    while (true)
+
     {
-        {
-            std::lock_guard<std::mutex> lock(data.bankMutex);
-            std::cout << "Client " << data.clientId << " checked balance of account " << data.accountNumber << ": " << data.account.getBalance() << std::endl;
-        }
-        std::this_thread::sleep_for(std::chrono::milliseconds(22000));
+        std::lock_guard<std::mutex> lock(data.bankMutex);
+        std::cout << "Client " << data.clientId << " checked balance of account " << data.accountNumber << ": " << data.account.getBalance() << std::endl;
     }
+    std::this_thread::sleep_for(std::chrono::milliseconds(2200));
 }
 
 void clientSimulation(Bank &bank, int clientId, std::mutex &bankMutex)
@@ -88,12 +79,27 @@ void clientSimulation(Bank &bank, int clientId, std::mutex &bankMutex)
     }
 
     RandomTransactionData data = {*account, clientId, accountNumber, amount, bankMutex};
+    while (running)
+    {
+        std::thread depositThread(depositRandom, data);
+        std::thread withdrawThread(withdrawRandom, data);
+        std::thread showBalanceThread(showBalance, data);
 
-    std::thread depositThread(depositRandom, data);
-    std::thread withdrawThread(withdrawRandom, data);
-    std::thread showBalanceThread(showBalance, data);
+        depositThread.join();
+        withdrawThread.join();
+        showBalanceThread.join();
+        std::unique_lock<std::mutex> lock(mtx);
+        cv.wait_for(lock, std::chrono::seconds(1), []
+                    { return !running; }); // HÄR VA VI SENAST, KOLLA HÄR MISSA MIG INTE
+    }
+}
 
-    depositThread.join();
-    withdrawThread.join();
-    showBalanceThread.join();
+void waitForEnter()
+{
+    std::cin.get();
+    {
+        std::lock_guard<std::mutex> lock(mtx);
+        running = false;
+    }
+    cv.notify_all();
 }
